fix(snake_case)

+++ src/services/post.js
@@ -17,24 +17,27 @@ export default class PostService {
 		}
 	}
 
-	async read(userId, postId) {
+	async read(user_id, post_id) {
 		try {
 			let post;
-			if (userId) {
+			if (user_id) {
 				post = await Post.scope([
-					{ name: "withAuthenticatedUser", options: userId },
-				]).findByPk(postId, {
+					{ name: "withAuthenticatedUser", options: user_id },
+				]).findByPk(post_id, {
 					attributes: [
 						"id",
-						"userId",
+						"user_id",
 						"title",
 						"content",
 						"total_likes",
-						"createdAt",
+						"created_at",
+						"updated_at",
+
+						
 					],
 				});
 			} else {
-				post = await Post.findByPk(postId);
+				post = await Post.findByPk(post_id);
 			}
 
 			if (!post) {
@@ -47,16 +50,16 @@ export default class PostService {
 		}
 	}
 
-	async readAll(userId, meta) {
+	async readAll(user_id, meta) {
 		try {
 			const scopes = [];
 			const Pagination = PaginationUtils.config({
 				page: meta.page,
 			});
-			if (userId) {
+			if (user_id) {
 				scopes.push({
 					name: "withAuthenticatedUser",
-					options: userId,
+					options: user_id,
 				});
 			}
 
@@ -65,13 +68,15 @@ export default class PostService {
 				raw: false,
 				attributes: [
 					"id",
-					"userId",
+					"user_id",
 					"title",
 					"content",
 					"total_likes",
-					"createdAt",
+					"created_at",
+					"updated_at",
+
 				],
-				order: [["createdAt", "DESC"]],
+				order: [["created_at", "DESC"]],
 			});
 
 			let totalItemsPromise;
@@ -93,25 +98,29 @@ export default class PostService {
 		}
 	}
 
-	async update(body, id) {
+	async update(body, id, user_id) {
 		const transaction = await Post.sequelize.transaction();
-
+	
 		try {
-			const [_, [updatedPost]] = await Post.update(body, {
-				where: { id },
+			const [affectedRows, [updatedPost]] = await Post.update(body, {
+				where: { id, user_id },
 				returning: true,
 				transaction,
 			});
-
+	
+			if (!updatedPost) {
+				throw new Error("Post not found or user not authorized to update this post");
+			}
+	
 			await transaction.commit();
 			return updatedPost;
 		} catch (error) {
 			await transaction.rollback();
 			throw error;
 		}
-	}
+	}	
 
-	async delete(id) {
+	async delete(id, user_id) {
 		try {
 			const post = await Post.findByPk(id);
 
@@ -119,7 +128,11 @@ export default class PostService {
 				throw new Error("Post not found");
 			}
 
-			await Post.destroy({ where: { id } });
+			const deletedPost = await Post.destroy({ where: { id, user_id } });
+
+			if (!deletedPost) {
+				throw new Error("User not authorized to delete this post");
+			}
 
 			return { message: "Post deleted successfully" };
 		} catch (error) {
